Add book class
